refactor : repository의존성 제거 및 도메인 서비스로 로직 분리
